Make parameter binding API public
This allows others to build proxies without using internal code.

Fixes #14554 #15252

Internal code is still being used for conventions.
