+++ src/EFCore/Metadata/IMutableServiceProperty.cs
@@ -1,6 +1,9 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+using JetBrains.Annotations;
+using Microsoft.EntityFrameworkCore.Metadata.Internal;
+
 namespace Microsoft.EntityFrameworkCore.Metadata
 {
     /// <summary>
@@ -19,5 +22,10 @@ namespace Microsoft.EntityFrameworkCore.Metadata
         ///     Gets the type that this property belongs to.
         /// </summary>
         new IMutableEntityType DeclaringEntityType { get; }
+
+        /// <summary>
+        ///    The <see cref="ServiceParameterBinding"/> for this property.
+        /// </summary>
+        new ServiceParameterBinding ParameterBinding { get; [param: CanBeNull] set; }
     }
 }
