--- src/EFCore/Metadata/Internal/ContextParameterBinding.cs
@@ -1,52 +1,0 @@
-// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using System;
-using System.Linq.Expressions;
-using JetBrains.Annotations;
-using Microsoft.EntityFrameworkCore.Storage;
-
-namespace Microsoft.EntityFrameworkCore.Metadata.Internal
-{
-    /// <summary>
-    ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-    ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-    ///     any release. You should only use it directly in your code with extreme caution and knowing that
-    ///     doing so can result in application failures when updating to a new Entity Framework Core release.
-    /// </summary>
-    public class ContextParameterBinding : ServiceParameterBinding
-    {
-        /// <summary>
-        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-        ///     any release. You should only use it directly in your code with extreme caution and knowing that
-        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
-        /// </summary>
-        public ContextParameterBinding(
-            [NotNull] Type contextType,
-            [CanBeNull] IPropertyBase consumedProperty = null)
-            : base(contextType, contextType, consumedProperty)
-        {
-        }
-
-        /// <summary>
-        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-        ///     any release. You should only use it directly in your code with extreme caution and knowing that
-        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
-        /// </summary>
-        public override Expression BindToParameter(
-            Expression materializationExpression,
-            Expression entityTypeExpression)
-        {
-            var propertyExpression
-                = Expression.Property(
-                    materializationExpression,
-                    MaterializationContext.ContextProperty);
-
-            return ServiceType != typeof(DbContext)
-                ? (Expression)Expression.TypeAs(propertyExpression, ServiceType)
-                : propertyExpression;
-        }
-    }
-}
